Prototype for #401

+++ src/PeriLab.jl
@@ -61,6 +61,7 @@ import .Logging_Module
 import .IO
 using .ModuleLoader
 using .Solver_Manager
+using .Parameter_Handling: get_initial_time
 
 PERILAB_VERSION = "2.2.5"
 
@@ -385,11 +386,12 @@ function run(filename::String;
                 output_dir = filedirectory
             end
 
-            if !reload
-                Data_Manager.initialize_data()
-            else
-                @info "PeriLab started in the reload mode"
+            Data_Manager.initialize_data()
+            if reload
+                @info "Reloading from checkpoint"
+                Data_Manager.read_checkpoint!(output_dir, rank)
             end
+
             Data_Manager.set_silent(silent)
             Data_Manager.set_verbose(verbose)
             @timeit "IO.initialize_data" params,
@@ -477,6 +479,11 @@ function run(filename::String;
                                                                                        IO.write_results,
                                                                                        silent)
                 end
+                if reload
+                    @info "Writing checkpoint"
+                    @timeit "Data_Manager.write_checkpoint" Data_Manager.write_checkpoint(output_dir,
+                                                                                          rank)
+                end
             end
 
         catch e
@@ -498,7 +505,7 @@ function run(filename::String;
                 IO.merge_exodus_files(result_files, output_dir)
             end
             MPI.Barrier(comm)
-            if (size > 1 && !debug) || dry_run
+            if (size > 1 && !debug && !reload) || dry_run
                 IO.delete_files(result_files, output_dir)
             end
         end
